Allow e2e tests to be run without service restart
Start e2e test services:

	./scripts/e2e-servers.js

Set up e2e expectations etc. for the first run of tests:

	grunt exec:deploytest

Run e2e tests without starting services:

	grunt protractor:tests_only
