+++ tests/protractor/tests-only.conf.js
@@ -1,0 +1,3 @@
+const BaseConfig = require('./base.conf.js');
+
+module.exports = new BaseConfig();
